flaskAPIService/requirements.txt
Flask
Flask-PyMongo
Flask-Redis
python-dotenv
